requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
pypdf
chromadb
sentence-transformers
google-generativeai
celery
redis
streamlit